modify

+++ app/Http/Controllers/EmpleadoController.php
@@ -47,7 +47,8 @@ class EmpleadoController extends Controller
 
         // con este metodo insertamos los datos en la base de datos
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
+        
+        return redirect('empleado')->with('mensaje','Empleado agregado con exito');
     }
 
     /**
@@ -105,8 +106,11 @@ class EmpleadoController extends Controller
      */
     public function destroy($id)
     {
-        // con este metodo eliminamos un registro de la base de datos
-        Empleado::destroy($id);
-         return redirect('empleado');
+        // con este metodo eliminamos lo que alla en la carpeta uploads, y datos en la base de datos
+        $empleado = Empleado::findOrFail($id);
+        if(Storage::delete('public/'.$empleado->foto)){
+            Empleado::destroy($id);
+        }
+         return redirect('empleado')->with('mensaje','Empleado eliminado con exito');
     }
 }
